Reorganize ETSI-CTI TITAN docker

+++ from-sources/build.sh
@@ -7,7 +7,10 @@ set -vx
 
 #check and build stfubuntu image
 DOCKER_ID=`docker ps -a | grep -e stfubuntu | awk '{ print $1 }'`
-docker rm --force ${DOCKER_ID}
+if [ ! -z "${DOCKER_ID}" ]
+then
+    docker rm --force ${DOCKER_ID}
+fi
 if [ -z `docker images -q stfubuntu` ]; then
   docker build --tag stfubuntu:18.04 -f Dockerfile.stfubuntu --force-rm  . || exit 1
 fi
